requirements.txt
diffusers==0.29.2
einops
numpy==1.26.3
Pillow==10.2.0
PyYAML
safetensors
tqdm
transformers==4.37.2
opencv-python
omegaconf
kornia
controlnet-aux==0.0.7
accelerate==0.26.1